api/requirements.txt
# 핵심 프레임워크
Django>=4.0
djangorestframework>=3.14.0

# 데이터베이스
mysqlclient>=2.1.0

# 환경 설정
python-dotenv>=0.19.2

# 웹 서버 & 배포
gunicorn>=21.0.0
whitenoise>=6.5.0

# Django 확장 기능
django-cors-headers>=3.13.0
django-filter>=22.1

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0

# 비동기 작업
celery>=5.3.0
redis>=4.5.0

# 웹 스크래핑
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 이미지 처리
Pillow>=10.0.0

# AI
scikit-learn>=1.5.0
xgboost>=3.0.0
imbalanced-learn>=0.12.0